refactor(codex): move spend wire schema beside its parser
